# internal/keyspan: modify FragmentIterator seek semantics