--- a/internal/keyspan/testdata/iter
+++ b/internal/keyspan/testdata/iter
@@ -8,19 +8,25 @@ iter
 seek-ge a
 seek-ge b
 seek-ge c
+seek-ge cat
 seek-ge d
 seek-lt a
 seek-lt b
 seek-lt c
+seek-lt cat
 seek-lt d
+seek-lt e
 ----
 a-b:{(#2,SET) (#1,SET)}
 b-c:{(#2,SET) (#1,SET)}
+c-d:{(#2,SET) (#1,SET)}
 c-d:{(#2,SET) (#1,SET)}
 .
 .
 a-b:{(#2,SET) (#1,SET)}
 b-c:{(#2,SET) (#1,SET)}
+c-d:{(#2,SET) (#1,SET)}
+c-d:{(#2,SET) (#1,SET)}
 c-d:{(#2,SET) (#1,SET)}
 
 iter